chore: split the sidebar items into separate component
