+++ src/components/SideBar.tsx
@@ -1,11 +1,12 @@
-import { ReactElement, JSXElementConstructor } from 'react'
-import { Link, useLocation } from 'react-router-dom'
+import type { ReactElement } from 'react'
+import { Link } from 'react-router-dom'
 
 import { createStyles, Theme, makeStyles } from '@material-ui/core/styles'
-import { ListItemText, ListItemIcon, ListItem, Divider, List, Drawer, Link as MUILink } from '@material-ui/core'
+import { ListItemText, ListItem, Divider, List, Drawer, Link as MUILink } from '@material-ui/core'
 import { OpenInNewSharp } from '@material-ui/icons'
 import { Home, FileText, DollarSign, Share2, Settings, Layers } from 'react-feather'
 import { ROUTES } from '../routes'
+import SideBarItem from './SideBarItem'
 
 import Logo from '../assets/logo.svg'
 
@@ -65,26 +66,6 @@ interface Props {
   isOk: boolean
 }
 
-interface PropsRow {
-  icon: ReactElement
-  path: string
-  label: string
-}
-
-function Row({ icon, path, label }: PropsRow): ReactElement {
-  const classes = useStyles()
-  const location = useLocation()
-
-  return (
-    <Link to={path} key={path} style={{ color: '#9f9f9f', textDecoration: 'none' }}>
-      <ListItem button selected={location.pathname === path} disableRipple>
-        <ListItemIcon className={location.pathname === path ? classes.activeSideBar : ''}>{icon}</ListItemIcon>
-        <ListItemText primary={label} />
-      </ListItem>
-    </Link>
-  )
-}
-
 export default function SideBar(props: Props): ReactElement {
   const classes = useStyles()
 
@@ -98,12 +79,14 @@ export default function SideBar(props: Props): ReactElement {
         </div>
         <List>
           {navBarItems.map(p => (
-            <Row
-              key={p.path}
-              icon={<p.icon style={{ height: '36px', color: '#9f9f9f' }} />}
-              path={p.path}
-              label={p.label}
-            />
+            <Link to={p.path} key={p.path} style={{ color: '#9f9f9f', textDecoration: 'none' }}>
+              <SideBarItem
+                key={p.path}
+                icon={<p.icon style={{ height: '36px', color: '#9f9f9f' }} />}
+                path={p.path}
+                label={p.label}
+              />
+            </Link>
           ))}
         </List>
         <Divider />
